Create github management Script

+++ GITHUB_MANAGEMENT/2_delete_branch.sh
@@ -8,14 +8,19 @@ export LC_ALL=C
 
 ### Git Repo Sync ###
 echo
-echo "[INFO] Github PR Branch Delete : pr_branch"
+echo "[WARNING] Be sure to go through the [ Create pull request ] and [ Merge pull request ] steps before executing."
+echo "[QUESTION] Do you want Execute ? : y or n"
+read ANSWER
+echo
 
-git pull origin main
-
-git checkout main
-
-git branch -D pr_branch
-
-git push --delete origin pr_branch
-
-git pull origin main
+if [ "${ANSWER}" == "y" ]
+then
+	echo "[INFO] Github PR Branch Delete : pr_branch"
+	git pull origin main
+	git checkout main
+	git branch -D pr_branch
+	git push --delete origin pr_branch
+	git pull origin main
+else
+	echo "[INFO] Github PR Branch Not Delete : pr_branch"
+fi
